fix: remove abstract class modifier
