+++ app/src/main/kotlin/model/DirectedGraph.kt
@@ -1,6 +1,83 @@
 package model
 
-import model.edges.DirectedEdge
-import model.internalGraphs._DirectedGraph
+import model.abstractGraph.Edge
+import model.abstractGraph.Graph
+import model.abstractGraph.Vertex
 
-class DirectedGraph<D> : _DirectedGraph<D, DirectedEdge<D>>()
+open class DirectedGraph<D> : Graph<D>() {
+    override fun addEdge(vertex1: Vertex<D>, vertex2: Vertex<D>): Edge<D> {
+        if (vertex1 == vertex2)
+            throw IllegalArgumentException("Can't add edge from vertex to itself.")
+
+        if (vertex1 !in adjacencyMap.keys || vertex2 !in adjacencyMap.keys)
+            throw NoSuchElementException("Vertex1 or vertex2 is not in the adjacency map.")
+
+        val newEdge = Edge(vertex1, vertex2) as Edge<D>
+        edges.add(newEdge)
+        adjacencyMap[vertex1]?.add(vertex2)
+
+        return newEdge
+    }
+
+    override fun removeEdge(edgeToRemove: Edge<D>): Edge<D> {
+        if (edgeToRemove !in edges)
+            throw NoSuchElementException("Edge is not in the graph")
+
+        val vertex1 = edgeToRemove.vertex1
+        val vertex2 = edgeToRemove.vertex2
+
+        adjacencyMap[vertex1]?.remove(vertex2)
+        edges.remove(edgeToRemove)
+
+        return edgeToRemove
+    }
+
+    fun findSCC(): ArrayList<ArrayList<Vertex<D>>> { // SCC - Strongly Connected Components (by Kosaraju)
+        val visited = mutableMapOf<Vertex<D>, Boolean>().withDefault { false }
+        val stack = ArrayDeque<Vertex<D>>()
+        val component = arrayListOf<Vertex<D>>()
+        val sccList: ArrayList<ArrayList<Vertex<D>>> = arrayListOf()
+
+        fun auxiliaryDFS(srcVertex: Vertex<D>, componentList: ArrayList<Vertex<D>>) {
+            visited[srcVertex] = true
+            componentList.add(srcVertex)
+            adjacencyMap[srcVertex]?.forEach { vertex2 ->
+                if (visited[vertex2] != null && visited[vertex2] != true) {
+                    auxiliaryDFS(vertex2, componentList)
+                }
+            }
+            stack.add(srcVertex)
+        }
+
+        for (vertex in adjacencyMap.keys) {
+            if (visited[vertex] != null && visited[vertex] != true) auxiliaryDFS(vertex, component)
+        }
+
+        reverseGraph()
+        visited.clear()
+        component.clear()
+
+        while (stack.isNotEmpty()) {
+            val vertex = stack.removeLast()
+            if (visited[vertex] != null && visited[vertex] != true) {
+                val currentComponent = arrayListOf<Vertex<D>>()
+                auxiliaryDFS(vertex, currentComponent)
+                sccList.add(currentComponent)
+            }
+        }
+        return sccList
+    }
+
+    private fun reverseGraph() {
+        val reversedAdjacencyMap = mutableMapOf<Vertex<D>, ArrayList<Vertex<D>>>()
+        for (vertex in adjacencyMap.keys) {
+            adjacencyMap[vertex]?.forEach { vertex2 ->
+                reversedAdjacencyMap[vertex2] = reversedAdjacencyMap[vertex2] ?: ArrayList()
+                reversedAdjacencyMap[vertex2]?.add(vertex)
+            }
+        }
+        adjacencyMap.clear()
+        adjacencyMap.putAll(reversedAdjacencyMap)
+    }
+}
+
