+++ app/src/main/kotlin/model/abstractGraph/Edge.kt
@@ -1,8 +1,6 @@
 package model.abstractGraph
 
-abstract class Edge<D> {
-    abstract val vertex1: Vertex<D>
-    abstract val vertex2: Vertex<D>
+class Edge<D>(val vertex1: Vertex<D>, val vertex2: Vertex<D>) {
 
     fun isIncident(vertex: Vertex<D>) = (vertex1 == vertex || vertex2 == vertex)
 }
