--- app/src/main/kotlin/model/edges/UndirectedEdge.kt
@@ -1,6 +1,0 @@
-package model.edges
-
-import model.abstractGraph.Edge
-import model.abstractGraph.Vertex
-
-open class UndirectedEdge<D>(override val vertex1: Vertex<D>, override val vertex2: Vertex<D>) : Edge<D>()
